% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDclassifier.R
\docType{package}
\name{LDclassifier}
\alias{LDclassifier}
\alias{LDclassifier-package}
\title{LDclassifier: Group chromosomes by recombination history}
\description{
LDclassifier infers recombination history based on a mixture model that, given a pair
of SNP-blocks, separates chromosomes in two populations, one with high Linkage Disequilibrium (LD)
and low recombination (linkage) and another with low LD and high recombination. The
method use the classification of several SNP-block pairs in a region to group chromosomes
in clusters with different recombination history. This package takes as input genotype phased data.
}
