% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newFreqBase.R
\name{newFreqBase}
\alias{newFreqBase}
\title{Reestimate recomb block frequencies}
\usage{
newFreqBase(Resp, Block)
}
\arguments{
\item{Resp}{Numerical with the responsibilities for recomb population}

\item{Block}{Character with the blocks genotypes}
}
\value{
Numerical with the new block frequencies for recomb population
}
\description{
Reestimate recomb block frequencies
}
