% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newFreqLD.R
\name{newFreqLD}
\alias{newFreqLD}
\title{Reestimate linkage block frequencies}
\usage{
newFreqLD(Resp, Block)
}
\arguments{
\item{Resp}{Numerical with the responsibilities for linkage population}

\item{Block}{Character with the blocks genotypes}
}
\value{
Numerical with the new block frequencies for linkage population
}
\description{
Reestimate linkage block frequencies
}
