% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLDclassifier.R
\name{runLDclassifier}
\alias{runLDclassifier}
\title{Main function to run all the LDclassifier pipeline}
\usage{
runLDclassifier(haplos, annot, clusters = 2, PCs = 1, ...)
}
\arguments{
\item{haplos}{Matrix with the haplotypes (SNPs in rows, samples in columns)}

\item{annot}{GenomicRanges with the SNPs annotation}

\item{clusters}{Numeric with the clusters used in k-means}

\item{PCs}{Numeric with the number of PCA components used to make the clustering.}

\item{...}{Further arguments passed to runLDmixtureModel}
}
\value{
A list with two elements:
\itemize{
 \item{"class"}{Cluster classification of the chromosomes}
 \item{"pc"}{Responsibilities PCA}
}
}
\description{
This function takes as input phased chromosomes and returns the responsibilities
PCA and the cluster classification of the samples.
}
