% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateModel.R
\name{updateModel}
\alias{updateModel}
\title{Update model parameters}
\usage{
updateModel(params, funcProps1, funcProps2)
}
\arguments{
\item{params}{List with current parameter values}

\item{funcProps1}{Function used to update block frequencies for population 1}

\item{funcProps2}{Function used to update block frequencies for population 2}
}
\value{
List with the updated frequencies
}
\description{
Update model parameters
}
